refactor: filter panel types
